// types/bwgr/character.d.ts
import type { BwgrAbilityId } from "./ability";
import type { BwgrLifepathId } from "./lifepath";
import type { BwgrQuestionId } from "./question";
import type { BwgrResourceId, BwgrResourceTypeId } from "./resource";
import type { BwgrSkillId } from "./skill";
import type { BwgrStockId } from "./stock";
import type { BwgrTraitId } from "./trait";


export type BwgrCharacterBurnerModals = "lp" | "randLp" | "re" | "geSk" | "geTr" | "qu" | "so";
export type BwgrShades = "B" | "G" | "W";

export interface BwgrCharacterAttribute {
  id: BwgrAbilityId;
  name: string;
  hasShade: boolean;
  shadeShifted: boolean;
  exponent: number;
}

export interface BwgrCharacterSkill {
  id: BwgrSkillId;
  name: string;
  type: "Mandatory" | "Lifepath" | "General";
  isSpecial: boolean;
  isOpen: "no" | "yes" | "double";
  advancement: { general: number; lifepath: number; };
}

export interface BwgrCharacterTrait {
  id: BwgrTraitId;
  name: string;
  type: "Mandatory" | "Lifepath" | "Common" | "General";
  isOpen: boolean;
}

export interface BwgrCharacterResource {
  id: BwgrResourceId;
  name: string;
  type: [id: BwgrResourceTypeId, name: string];
  modifiers: string[];
  cost: number;
  description: string;
}

export interface BwgrCharacterSpecialStock {
  brutalLifeTraits: ([id: BwgrTraitId, name: string] | "No Trait" | undefined)[];
  huntingGround: undefined | BwgrHuntingGroundsList;
}

export interface BwgrCharacterSpecial {
  stock: BwgrCharacterSpecialStock;
  companionLifepath: Record<string, BwgrLifepathId>;
  variableAge: Record<BwgrLifepathId, number>;
  companionSkills: Record<string, BwgrSkillId[]>;
  chosenSubskills: Record<BwgrSkillId, BwgrSkillId[]>;
}

export interface BwgrCharacterStockLimits {
  beliefs: number;
  instincts: number;
  stats: {
    [key: string]: { min: number; max: number; };
    Will: { min: number; max: number; };
    Perception: { min: number; max: number; };
    Power: { min: number; max: number; };
    Agility: { min: number; max: number; };
    Forte: { min: number; max: number; };
    Speed: { min: number; max: number; };
  };
  attributes: number;
}

export interface BwgrStatData {
  poolType: "Mental" | "Physical";
  shadeShifted: boolean;
  mainPoolSpent: { shade: number; exponent: number; };
  eitherPoolSpent: { shade: number; exponent: number; };
}

export interface BwgrCharacterQuestion {
  id: BwgrQuestionId;
  name: string;
  question: string;
  answer: boolean;
}

export type BwgrCharacterTraitEffect =
  | { roundUp: BwgrAbilityId; }
  | { roundUp: "Mortal Wound"; }
  | { callOn: BwgrAbilityId; }
  | { callOn: BwgrSkillId; };

export type BwgrHuntingGroundsList = "Waste" | "Marginal" | "Typical" | "Plentiful" | "Untouched";

export interface BwgrPoints {
  total: number;
  spent: number;
  remaining: number;
}

export interface BwgrAbilityPoints {
  shade: BwgrShades;
  exponent: number;
}

export interface BwgrBurningCharacter {
  basics: {
    name: string;
    concept: string;
    gender: "Male" | "Female";
    stock: [id: BwgrStockId, name: string];
    beliefs: { name: string; belief: string; }[];
    instincts: { name: string; instinct: string; }[];
  };
  lifepaths: {
    lifepaths: BwgrLifepath[];
  };
  stats: {
    stats: Record<string, BwgrStatData>;
  };
  skills: {
    skills: BwgrCharacterSkill[];
  };
  traits: {
    traits: BwgrCharacterTrait[];
  };
  attributes: {
    attributes: BwgrCharacterAttribute[];
  };
  resources: {
    resources: Record<Guid, BwgrCharacterResource>;
  };
  misc: {
    special: BwgrCharacterSpecial;
    questions: BwgrCharacterQuestion[];
    limits: BwgrCharacterStockLimits;
    traitEffects: BwgrCharacterTraitEffect[];
  };
}